pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dudoxx-whisper-stt"
version = "1.1.0"
description = "Dudoxx Real-time Speech-to-Text - Ultra-low-latency transcription for healthcare (based on WhisperLiveKit)"
readme = "README.md"
authors = [
    { name = "Dudoxx UG" },
    { name = "Quentin Fuxa", email = "original@whisperlivekit.dev" }
]
maintainers = [
    { name = "Walid Boudabbous", email = "w.boudabbous@dudoxx.com" }
]
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Medical Science Apps."
]
dependencies = [
    "fastapi",
    "librosa",
    "soundfile>=0.13.1",
    "uvicorn",
    "websockets",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "huggingface-hub>=1.0.0",
    "tqdm",
    "tiktoken",
    'triton>=2.0.0; platform_machine == "x86_64" and (sys_platform == "linux" or sys_platform == "linux2")'
]

[project.optional-dependencies]
translation = ["nllw"]
sentence_tokenizer = ["mosestokenizer", "wtpsplit"]
macos = ["mlx-whisper>=0.4.0"]
diarization = [
    "pyannote.audio>=4.0.0",
    "diart>=0.9.0",
    "rx>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/Dudoxx/dudoxx-whisper-stt"
Documentation = "https://github.com/Dudoxx/dudoxx-whisper-stt/blob/main/docs/"
Repository = "https://github.com/Dudoxx/dudoxx-whisper-stt"
"Original Project" = "https://github.com/QuentinFuxa/WhisperLiveKit"

[project.scripts]
dudoxx-stt = "whisperlivekit.basic_server:main"
dudoxx-whisper = "whisperlivekit.basic_server:main"
# Keep original commands for compatibility
whisperlivekit-server = "whisperlivekit.basic_server:main"
wlk = "whisperlivekit.basic_server:main"

[tool.setuptools]
packages = [
    "whisperlivekit",
    "whisperlivekit.diarization",
    "whisperlivekit.simul_whisper",
    "whisperlivekit.whisper",
    "whisperlivekit.whisper.assets",
    "whisperlivekit.whisper.normalizers",
    "whisperlivekit.web",
    "whisperlivekit.local_agreement",
    "whisperlivekit.silero_vad_models"
]

[tool.setuptools.package-data]
whisperlivekit = ["web/*.html", "web/*.css", "web/*.js", "web/src/*.svg"]
"whisperlivekit.whisper.assets" = ["*.tiktoken", "*.npz"]
"whisperlivekit.silero_vad_models" = ["*.jit", "*.onnx"]
